+++ salalib/segmmodules/segmmetric.cpp
@@ -136,7 +136,7 @@ bool SegmentMetric::run(Communicator *comm, const Options &options, ShapeGraph &
 
                 connected_cursor = iter->first.ref;
 
-                if (seen[connected_cursor] > segdepth && connected_cursor != cursor) {
+                if (seen[connected_cursor] > segdepth && static_cast<size_t>(connected_cursor) != cursor) {
                     bool seenalready = (seen[connected_cursor] == 0xffffffff) ? false : true;
                     float length = seglengths[connected_cursor];
                     audittrail[connected_cursor] =
